requirements.txt
python-dotenv==1.0.0
ccxt==4.0.0
aiohttp==3.8.5
asyncio==3.4.3
